Move PostHistory create and pop functions  to PostHistory.go
