+++ application/models/post_fixtures_test.go
@@ -436,114 +436,3 @@ func createFixturesForGetLocationForNotifications(ms *ModelSuite) PostFixtures {
 		Locations: locations,
 	}
 }
-
-func createFixturesForTestPost_createNewHistory(ms *ModelSuite) PostFixtures {
-	org := Organization{AuthConfig: "{}", Uuid: domain.GetUuid()}
-	createFixture(ms, &org)
-
-	unique := org.Uuid.String()
-	users := Users{
-		{Email: unique + "_user0@example.com", Nickname: unique + "User0", Uuid: domain.GetUuid()},
-		{Email: unique + "_user1@example.com", Nickname: unique + "User1", Uuid: domain.GetUuid()},
-	}
-	for i := range users {
-		createFixture(ms, &users[i])
-	}
-
-	posts := Posts{
-		{Title: "Post1 Title"},
-		{Title: "Post2 Title"},
-	}
-	locations := make(Locations, len(posts))
-	for i := range posts {
-		locations[i].Description = "location " + strconv.Itoa(i)
-		createFixture(ms, &locations[i])
-
-		posts[i].Uuid = domain.GetUuid()
-		posts[i].Status = PostStatusOpen
-		posts[i].Type = "type"
-		posts[i].Size = PostSizeTiny
-		posts[i].CreatedByID = users[0].ID
-		posts[i].OrganizationID = org.ID
-		posts[i].DestinationID = locations[i].ID
-		posts[i].ReceiverID = nulls.NewInt(users[i].ID)
-		createFixture(ms, &posts[i])
-	}
-
-	pHistory := PostHistory{
-		Status:     PostStatusOpen,
-		PostID:     posts[0].ID,
-		ReceiverID: nulls.NewInt(posts[0].CreatedByID),
-	}
-	createFixture(ms, &pHistory)
-
-	return PostFixtures{
-		Users:         users,
-		Posts:         posts,
-		PostHistories: PostHistories{pHistory},
-	}
-}
-
-func createFixturesForTestPost_popHistory(ms *ModelSuite) PostFixtures {
-	org := Organization{AuthConfig: "{}", Uuid: domain.GetUuid()}
-	createFixture(ms, &org)
-
-	unique := org.Uuid.String()
-	users := Users{
-		{Email: unique + "_user0@example.com", Nickname: unique + "User0", Uuid: domain.GetUuid()},
-		{Email: unique + "_user1@example.com", Nickname: unique + "User1", Uuid: domain.GetUuid()},
-	}
-	for i := range users {
-		createFixture(ms, &users[i])
-	}
-
-	posts := Posts{
-		{Title: "Post1 Title", ProviderID: nulls.NewInt(users[1].ID)},
-		{Title: "Post2 Title"},
-	}
-	locations := make(Locations, len(posts))
-	for i := range posts {
-		locations[i].Description = "location " + strconv.Itoa(i)
-		createFixture(ms, &locations[i])
-
-		posts[i].Uuid = domain.GetUuid()
-		posts[i].Status = PostStatusAccepted
-		posts[i].Type = "type"
-		posts[i].Size = PostSizeTiny
-		posts[i].CreatedByID = users[0].ID
-		posts[i].OrganizationID = org.ID
-		posts[i].DestinationID = locations[i].ID
-		posts[i].ReceiverID = nulls.NewInt(users[i].ID)
-		createFixture(ms, &posts[i])
-	}
-
-	pHistories := PostHistories{
-		{
-			Status:     PostStatusOpen,
-			PostID:     posts[0].ID,
-			ReceiverID: nulls.NewInt(posts[0].CreatedByID),
-		},
-		{
-			Status:     PostStatusCommitted,
-			PostID:     posts[0].ID,
-			ReceiverID: nulls.NewInt(posts[0].CreatedByID),
-			ProviderID: nulls.NewInt(users[1].ID),
-		},
-		{
-			Status:     PostStatusAccepted,
-			PostID:     posts[0].ID,
-			ReceiverID: nulls.NewInt(posts[0].CreatedByID),
-			ProviderID: nulls.NewInt(users[1].ID),
-		},
-	}
-
-	for i := range pHistories {
-		createFixture(ms, &pHistories[i])
-	}
-
-	return PostFixtures{
-		Users:         users,
-		Posts:         posts,
-		PostHistories: pHistories,
-	}
-}
